got some issues on collaboration feature

+++ src/services/postgres/PlaylistService.js
@@ -37,7 +37,7 @@ class PlaylistsService {
   }
 
   async deletePlaylistById(id, username) {
-    this.verifyPlaylistOwner(username);
+    await this.verifyPlaylistAccess(id, username);
     const query = {
       text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
       values: [id],
@@ -50,37 +50,47 @@ class PlaylistsService {
     }
   }
 
-  async verifyPlaylistOwner(username) {
+  async verifyPlaylistOwner(username, playistId) {
     const query = {
-      text: 'SELECT * FROM playlists WHERE id = $1',
-      values: [username],
+      text: 'SELECT * FROM playlists',
     };
 
     const res = await this._pool.query(query);
-    if (!res.rowCount) throw new InvariantError('Playlist tidak ditemukan');
-    const playlist = res.rows[0];
+
+    const idArr = res.rows.map((p) => ({
+      id: p.id,
+    }));
+    const usernameArr = res.rows.map((p) => ({
+      username: p.username,
+    }));
+
+    const matchingId = idArr.filter((p) => p.id == playistId).length;
+    const matchingUsername = usernameArr.filter((p) => p.username == username).length;
+    console.log(usernameArr, username);
 
     const errMsg = 'Anda tidak berhak mengakses resource ini';
-    if (playlist.username != username) throw new AuthorizationError(errMsg);
+    if (matchingId < 1) throw new NotFoundError('Playlist tidak ditemukan');
+    if (matchingUsername < 1) throw new AuthorizationError(errMsg);
   }
 
   async verifyPlaylistAccess(playlistId, userId) {
     try {
-      await this.verifyPlaylistOwner(playlistId, userId);
+      await this.verifyPlaylistOwner(userId, playlistId);
     } catch (error) {
       if (error instanceof NotFoundError) {
         throw error;
       }
       try {
-        await this._collaborationService.verifyCollaborator(playlistId, userId);
+        await this._collaborationService.verifyPlaylistCollaborator(playlistId, userId);
       } catch (error) {
         throw error;
       }
     }
   }
 
-  async addSongToPlaylist(playlistId, songId) {
-    this._songService.verifySongId(songId);
+  async addSongToPlaylist(playlistId, songId, owner) {
+    await this._songService.verifySongId(songId, owner);
+    await this.verifyPlaylistAccess(playlistId, owner);
 
     const id = `track-${nanoid(16)}`;
     const query = {
@@ -96,20 +106,26 @@ class PlaylistsService {
   }
 
   async getListOfSongs(username, playlistId) {
+    await this.verifyPlaylistAccess(playlistId, username);
     const query = {
-      text: `select playlists.id, name, username, songs.id, songs.title, songs.performer from 
-      playlists left join playlist_songs on playlists.id = playlist_songs.playlist_i
-      d left join songs on playlist_songs.song_id = songs.id WHERE username = $1 AND playlist.id = $2;
+      text: `select name, users.username, songs.id, songs.title, songs.performer from 
+      playlists left join playlist_songs on playlists.id = playlist_songs.playlist_id 
+      left join songs on playlist_songs.song_id = songs.id 
+      left join users on users.id = $1
+      WHERE users.id = $1 AND playlists.id = $2;
       `,
       values: [username, playlistId],
     };
 
     const res = await this._pool.query(query);
+
     if (!res.rowCount) throw new InvariantError('Playlist tidak valid');
+    return res.rows;
   }
 
-  async deleteSongFromPlaylist(playlistId, songId) {
-    this._songService.verifySongId(id);
+  async deleteSongFromPlaylist(playlistId, songId, owner) {
+    await this._songService.validateSongPayload(songId);
+    await this.verifyPlaylistAccess(playlistId, owner);
 
     const query = {
       text: 'DELETE FROM playlist_songs WHERE playlist_id = $1 AND song_id = $2',
